VIDEO: Renamed srt_parser.* to subtitles.*
