+++ video/subtitles.cpp
@@ -21,7 +21,7 @@
 
 #include "common/file.h"
 
-#include "video/srt_parser.h"
+#include "video/subtitles.h"
 
 namespace Video {
